permisos agregados a /config

+++ js/slash/config/edited.js
@@ -1,7 +1,12 @@
 const db = require('firebase-admin').firestore();
+const { Permissions } = require('discord.js');
 module.exports = {
     channel: async (client, interact, options) => {
         const lang = client.util.lang({ lang: client.servers.get(interact.guildId).lang, route: 'slash/config' }).edited.reset;
+        if(!interact.member.permissions.has([Permissions.FLAGS.MANAGE_MESSAGES, Permissions.FLAGS.MANAGE_CHANNELS])) return interact.reply({
+            content: lang.permissions,
+            ephemeral: true
+        });
         const channel = options.get('channel');
         db.collection(channel.channel.guild.id).doc("edited").update({ 
             channel: channel.value
@@ -17,6 +22,10 @@ module.exports = {
     }, 
     reset: async (client, interact, options) => {
         const lang = client.util.lang({ lang: client.servers.get(interact.guildId).lang, route: 'slash/config' }).edited.reset;
+        if(!interact.member.permissions.has([Permissions.FLAGS.MANAGE_MESSAGES, Permissions.FLAGS.MANAGE_CHANNELS])) return interact.reply({
+            content: lang.permissions,
+            ephemeral: true
+        });
         db.collection(interact.guildId).doc("edited").delete()
         interact.reply({
             content: lang.reply,
